Add fix for search projects, createProjectInV, createVersion

+++ Work/storage/project.js
@@ -1,27 +1,4 @@
 module.exports.project = {
-    /*getProject: function(conn, email){
-        let rows = conn.querySync(`SELECT project.file, datecreation, author FROM project JOIN users ON author=users.id WHERE email = '${email}'`);
-        if (rows.length > 0) {
-            console.log(rows.length);
-            return {
-                project: {
-                    file: rows[0].file,
-                    datecreate: rows[0].datecreation,
-                    datemodified: rows[0].datelastmodified,
-                    author: rows[0].author,
-                    depth: rows[i].depth,
-                },
-                error: null
-            };
-        }
-
-        return {
-            project: null,
-            error: "You haven`t any project"
-        };
-
-    },*/
-
     getProjectsCount: function (conn) {
         let rows = conn.querySync(`SELECT count(id) as count FROM project`);
         return rows[0].count
@@ -81,15 +58,7 @@ module.exports.project = {
         let row = conn.querySync(`SELECT id FROM users WHERE email = '${email}'`);
         if(row.length > 0){
             let idUs = row[0].id;
-            let res = conn.querySync(`SELECT file FROM project WHERE author = '${idUs}'`);
-            let samename = false;
-            if(res.length > 0){
-                for (let i = 0; i < res.length; i++) {
-                    if(res[i].file === project.file){
-                        samename = true;
-                    }
-                }
-            }
+            let samename = this.checkName(conn, idUs, project);
             if(samename) {
                return {
                    message: null,
@@ -113,45 +82,48 @@ module.exports.project = {
     },
 
     deleteProject: function (conn, file) {
-        let row = conn.querySync(`SELECT version.id FROM version JOIN project ON proot = project.id WHERE file = '${file}' `);
+        let row = conn.querySync(`SELECT version.id FROM version WHERE proot = '${file}' `);
         if(row.length > 0){
             for (let i = 0; i < row.length; i++) {
                 resId = row[i].id;
                 conn.querySync(`DELETE FROM version WHERE id = '${resId}'`);
             }
         }
-        conn.querySync(`DELETE FROM project WHERE file = '${file}'`);
+        conn.querySync(`DELETE FROM project WHERE id = '${file}'`);
     },
+
    createProjectInV: function(conn, email, project){
     let row = conn.querySync(`SELECT id FROM users WHERE email = '${email}'`);
     if(row.length > 0){
         let idUs = row[0].id;
-        let res = conn.querySync(`SELECT file FROM project WHERE author = '${idUs}'`);
-        let samename = false;
-        if(res.length > 0){
-            for (let i = 0; i < res.length; i++) {
-                if(res[i].file === project.file){
-                    samename = true;
-                }
-            }
-        }
+        let samename = this.checkName(conn, idUs, project);
         if(samename) {
            return {
                message: null,
                error: "Проект с данным именем уже существует. Пожалуйста, выберете другое имя"
             }
         } else {
-            let row = conn.querySync(`SELECT data, depth FROM version JOIN project ON proot = project.id WHERE version.id = ${project.rootver}`);
+            let row = conn.querySync(`SELECT version, data, depth FROM version JOIN project ON proot = project.id WHERE version.id = ${project.rootver}`);
             if(row.length > 0){  
                 conn.querySync(`INSERT into project(file, datecreation, datelastmodified, depth, author)` + 
                 `VALUES('${project.file}','${project.datecreate}','${project.datemodified}',${row[0].depth}, ${idUs});`);
-                let res = conn.querySync(`SELECT id FROM project WHERE file ='${project.file}'`);
-                conn.querySync(`INSERT into version(version, datecreation, datemodified, data, proot, authorv)` + 
-                    `VALUES('version 1', '${project.datecreate}', '${project.datecreate}', '${row[0].data}', ${res[0].id}, ${idUs});`);
-                return {
-                    message: "Проект успешно создан",
-                    error: null
+                let res = conn.querySync(`SELECT id FROM project WHERE file ='${project.file}' AND author = ${idUs}`);
+                
+                if(res.length > 0){
+                    conn.querySync(`INSERT into version(version, datecreation, datemodified, data, proot, authorv)` + 
+                    `VALUES('${row[0].version}', '${project.datecreate}', '${project.datecreate}', '${row[0].data}', ${res[0].id}, ${idUs});`);
+                    return {
+                        message: "Проект успешно создан",
+                        error: null
+                        }
+                }else {
+                    conn.querySync(`DELETE FROM project WHERE file = '${project.file} AND author = ${idUs}'`);
+                    return {
+                        message: null,
+                        error: "Не удалось создать проект"
+                     }
                 }
+               
             }
         }
     } else { 
@@ -160,23 +132,40 @@ module.exports.project = {
             error: "Пользователь не найден"
          }
     }
-    
    },
+
+   checkName: function(conn, idUs, project){
+    let res = conn.querySync(`SELECT file FROM project WHERE author = '${idUs}'`);
+    let samename = false;
+    if(res.length > 0){
+        for (let i = 0; i < res.length; i++) {
+            if(res[i].file === project.file){
+                samename = true;
+            }
+        }
+    } 
+    return samename;
+    },
 
    addTag: function(conn, tag){
     let row = conn.querySync(`SELECT authorv, proot FROM version WHERE id=${tag.rootver}`);
     if(row.length > 0){
         conn.querySync(`INSERT INTO tag(description) VALUES('${tag.description}') ON CONFLICT(description) DO NOTHING`);
         let res = conn.querySync(`SELECT id FROM tag WHERE description = '${tag.description}'`);
-        conn.querySync(`INSERT INTO project_tag(tag, project) VALUES(${res[0].id}, ${row[0].proot});`);
-        return {
-            error: null,
-            message: "Тег добавлен"
-        }   
+        if(res.length > 0){
+            conn.querySync(`INSERT INTO project_tag(tag, project) VALUES(${res[0].id}, ${row[0].proot});`);
+            return {
+                error: null
+            }   
+        } else {
+            return {
+                error: "Что-то пошло не так",
+            }
+        }
+       
     } else {
         return {
             error: "Что-то пошло не так",
-            message: null
         }
     }
    },
@@ -188,7 +177,6 @@ module.exports.project = {
            for(let i=0; i < row.length; i++){
             conn.querySync(`DELETE FROM tag WHERE id=${row[i].id}`);
            }
-           
        }
    },
 
@@ -219,14 +207,16 @@ module.exports.project = {
             return result;
         }
 
-        result.error = "Can`t get search information";
-        return result
+        result.error = "Не удалось обработать информацию для поиска";
+        return result;
    },
 
    searchProjects: function (conn, searchData, email) {
-    let startQuery = `SELECT DISTINCT p.id, file, datecreation, datelastmodified, users.name, users.email, depth FROM project as p JOIN users ON author=users.id `;
+    let startQuery = ``;
     let count = 0;
     let filters = searchData;
+    let flag = false;
+
     for(let key in searchData){
         if(searchData[key] != "") {
             count++;
@@ -234,10 +224,6 @@ module.exports.project = {
             delete filters[key];
         }
     }
-    // console.log(count);
-    // console.log(filters);
-    let flag = false;
-    
     
     keys = Object.keys(filters);
     if (typeof filters['tag'] !== "undefined") {
@@ -245,81 +231,26 @@ module.exports.project = {
     }
     if(flag){
         startQuery = `SELECT p.id, file, datecreation, datelastmodified, users.name, users.email, depth, description FROM project as p JOIN users ON author=users.id JOIN project_tag ON p.id = project_tag.project JOIN tag ON project_tag.tag=tag.id `; 
+    } else {
+        startQuery = `SELECT DISTINCT p.id, file, datecreation, datelastmodified, users.name, users.email, depth FROM project as p JOIN users ON author=users.id `;
     }
 
     let fieldName;
     let endQuery = ``;
     let lastQuery = ``;
     let orderQuery = ``;
-    switch(count){
-        case 1: {
-            fieldName = this.searchKey(keys, filters, 0);
-            lastQuery = this.searchInstruction(fieldName, filters);
-            endQuery = startQuery + `WHERE` + lastQuery;
-            break;
+
+    endQuery += startQuery + `WHERE`;
+
+    for(let i = 1; i <= count; i++ ) {
+        fieldName = this.searchKey(keys, i - 1);
+        lastQuery = this.searchInstruction(fieldName, filters);
+        endQuery += lastQuery;
+        if(count > 1 && i != count){
+            endQuery += ` AND`;
         }
-        case 2: {
-            fieldName1 = this.searchKey(keys, filters, 0);
-            lastQuery1 = this.searchInstruction(fieldName1, filters);
-            fieldName2 = this.searchKey(keys, filters, 1);
-            lastQuery2 = this.searchInstruction(fieldName2, filters);
-            endQuery = startQuery + `WHERE` + lastQuery1 + ` AND` + lastQuery2;
-            break;
-        }
-        case 3: {
-            fieldName1 = this.searchKey(keys, filters, 0);
-            lastQuery1 = this.searchInstruction(fieldName1, filters);
-            fieldName2 = this.searchKey(keys, filters, 1);
-            lastQuery2 = this.searchInstruction(fieldName2, filters);
-            fieldName3 = this.searchKey(keys, filters, 2);
-            lastQuery3 = this.searchInstruction(fieldName3, filters);
-            endQuery = startQuery + `WHERE` + lastQuery1 + ` AND` + lastQuery2 + ` AND` + lastQuery3;
-            break;
-        }
-        case 4: {
-            fieldName1 = this.searchKey(keys, filters, 0);
-            lastQuery1 = this.searchInstruction(fieldName1, filters);
-            fieldName2 = this.searchKey(keys, filters, 1);
-            lastQuery2 = this.searchInstruction(fieldName2, filters);
-            fieldName3 = this.searchKey(keys, filters, 2);
-            lastQuery3 = this.searchInstruction(fieldName3, filters);
-            fieldName4 = this.searchKey(keys, filters, 3);
-            lastQuery4 = this.searchInstruction(fieldName4, filters);
-            endQuery = startQuery + `WHERE` + lastQuery1 + ` AND` + lastQuery2 + ` AND` + lastQuery3 + ` AND` + lastQuery4;
-            break;
-        }
-        case 5: {
-            fieldName1 = this.searchKey(keys, filters, 0);
-            lastQuery1 = this.searchInstruction(fieldName1, filters);
-            fieldName2 = this.searchKey(keys, filters, 1);
-            lastQuery2 = this.searchInstruction(fieldName2, filters);
-            fieldName3 = this.searchKey(keys, filters, 2);
-            lastQuery3 = this.searchInstruction(fieldName3, filters);
-            fieldName4 = this.searchKey(keys, filters, 3);
-            lastQuery4 = this.searchInstruction(fieldName4, filters);
-            fieldName5 = this.searchKey(keys, filters, 4);
-            lastQuery5 = this.searchInstruction(fieldName4, filters);
-            endQuery = startQuery + `WHERE` + lastQuery1 + ` AND` + lastQuery2 + ` AND` + lastQuery3 + ` AND` + lastQuery4 + ` AND` + lastQuery5;
-            break;
-        }
-        case 6: {
-            fieldName1 = this.searchKey(keys, filters, 0);
-            lastQuery1 = this.searchInstruction(fieldName1, filters);
-            fieldName2 = this.searchKey(keys, filters, 1);
-            lastQuery2 = this.searchInstruction(fieldName2, filters);
-            fieldName3 = this.searchKey(keys, filters, 2);
-            lastQuery3 = this.searchInstruction(fieldName3, filters);
-            fieldName4 = this.searchKey(keys, filters, 3);
-            lastQuery4 = this.searchInstruction(fieldName4, filters);
-            fieldName5 = this.searchKey(keys, filters, 4);
-            lastQuery5 = this.searchInstruction(fieldName4, filters);
-            fieldName6 = this.searchKey(keys, filters, 5);
-            lastQuery6 = this.searchInstruction(fieldName4, filters);
-            endQuery = startQuery + `WHERE` + lastQuery1 + ` AND` + lastQuery2 + ` AND` + lastQuery3 + ` AND` + lastQuery4 + ` AND` + lastQuery5 + ` AND` + lastQuery6;
-            break;
-        }
-        
     }
+  
     console.log(endQuery);
     
     if(flag) {
@@ -338,8 +269,6 @@ module.exports.project = {
     if (rows.length > 0) {
         for (let i = 0, j = 1; i < rows.length; i++, j++) {
             let check = (rows[i].email === email)?true:false;
-            // console.log("check ", check);
-            // console.log(rows[i]);
             if(flag){
                 let capt = false;
                 if(i != rows.length-1 && rows[i].id == rows[j].id) {
@@ -386,7 +315,7 @@ module.exports.project = {
     result.error = "Проекты не найдены. Попробуйте изменить параметры поиска";
     return result
 },
-    searchKey: function (keys, filters, i){
+    searchKey: function (keys, i){
             switch(keys[i]){
                 case "fileName": 
                     return "file";
@@ -411,7 +340,6 @@ module.exports.project = {
         switch(fieldName){
             case "file": {
                 let searchName = filters["fileName"] + "%";
-                // console.log(searchName);
                 return ` ${fieldName} LIKE '${searchName}'`;
             }
             case "author": {
@@ -437,8 +365,6 @@ module.exports.project = {
                 if(filters["tag"].includes(" ")){
                     let str = filters["tag"];
                     let arrayOfStrings = str.split(' ');
-                    // console.log(arrayOfStrings);
-                    // console.log("length",arrayOfStrings.length);
                     let query = ``;
                     for(let i = 0; i < arrayOfStrings.length; i++){
                         query += ` ${fieldName} = '${arrayOfStrings[i]}'`;
@@ -452,5 +378,7 @@ module.exports.project = {
                 }
             }
         }
-    }
+    },
+
+   
 }
